chore(resource): adjusted type definitions of resources
